Deal with version update checks that will probably fail very often
We support looking up versions one level deeper than the existing version that is specified. If version is '1.2', we also look for versions '1.2.+'.
This versioning scheme is not used often, so will mostly result in errors while resolving. We don't want to bother the user with these errors,
so we mark them as allowed to fail silently and swallow them.