Catch case where buildSrc build fails
When buildSrc fails to build, the root project is never created,
resulting in an IllegalStateException when we try to get the rootProject
on buildFinished. This points the users towards a seeming failure of
this plugin, while actually their buildSrc code has failed. Suppress the
exception and just don't run an update check in this case.

Fixes #5