Resolve configurations to determine plugin updates in a task.
Gradle 7.0 drops support for resolving configurations after a project has finished. So we need to make sure that we resolve our configurations
while the project is executing. The easiest way to do this is in a task. To ensure that this task always runs, we register it as a finalizer for all other tasks.
A nice additional advantage of registering it as finalizer of all other tasks is that it runs completely at the end.

Because we want plugin update information as late as possible, and for all projects together, we collect all updates on a property of the task and print them after the build has completed.